Extract part of ReadyToRunCodegenNodeFactory to separate class
